Add basic Bolt 5.1 support
