+++ tests/unit/async_/test_driver.py
@@ -493,8 +493,7 @@ async def test_driver_factory_with_notification_filters(
         pool_cls, "open",
         return_value=mocker.AsyncMock(spec=pool_cls)
     )
-    if pool_cls is AsyncBoltPool:
-        open_mock.return_value.address = mocker.Mock()
+    open_mock.return_value.address = mocker.Mock()
     mocker.patch.object(AsyncBoltPool, "open", new=open_mock)
 
     if filters is ...:
@@ -536,8 +535,7 @@ async def test_session_factory_with_notification_filter(
     pool_cls = AsyncNeo4jPool if uri.startswith("neo4j://") else AsyncBoltPool
     pool_mock: t.Any = mocker.AsyncMock(spec=pool_cls)
     mocker.patch.object(pool_cls, "open", return_value=pool_mock)
-    if pool_cls is AsyncBoltPool:
-        pool_mock.address = mocker.Mock()
+    pool_mock.address = mocker.Mock()
     driver_filters = object()
     pool_mock.pool_config = PoolConfig(notification_filters=driver_filters)
     session_cls_mock = mocker.patch("neo4j._async.driver.AsyncSession",
