+++ tests/unit/common/work/test_summary.py
@@ -306,6 +306,7 @@ def test_summary_result_counters(summary_args_kwargs, counters_set) -> None:
     ((4, 4), "t_first"),
     ((5, 0), "t_first"),
     ((5, 1), "t_first"),
+    ((5, 2), "t_first"),
 ))
 def test_summary_result_available_after(
     summary_args_kwargs, exists, bolt_version, meta_name
@@ -334,6 +335,7 @@ def test_summary_result_available_after(
     ((4, 4), "t_last"),
     ((5, 0), "t_last"),
     ((5, 1), "t_last"),
+    ((5, 2), "t_last"),
 ))
 def test_summary_result_consumed_after(
     summary_args_kwargs, exists, bolt_version, meta_name
